Fix 64-bit warnings on release branch.

+++ groups/bdl/bdlt/bdlt_calendar.h
@@ -1791,7 +1791,7 @@ const Date& Calendar::lastDate() const
 inline
 int Calendar::length() const
 {
-    return d_nonBusinessDays.length();
+    return static_cast<int>(d_nonBusinessDays.length());
 }
 
 inline
